Adjusted reset callbacks functionality
The modification allows reset callbacks to be used
in a similar fashion as migrations.
When a reset callback is invoked, the execution will
be logged and consecutive executions will be suppressed.
The callbacks are limited to one unique class/method
combination when invoked on a particular instance of
a given class.